# rimind-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Rimind Backend API
  version: 1.0.0
  description: >
    RIMIND — AI-enabled maternal and child health support system API.
    All endpoints follow the SRS specification.
servers:
  - url: http://localhost:4000/api
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ExercisePlan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        stage:
          type: string
          enum: [PRENATAL, POSTNATAL]
        title:
          type: string
        description:
          type: string
        duration:
          type: integer
        source:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /exercise/{stage}:
    get:
      summary: Get exercise plans by stage
      parameters:
        - name: stage
          in: path
          required: true
          schema:
            type: string
            enum: [PRENATAL, POSTNATAL]
      responses:
        '200':
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExercisePlan'
  /exercise:
    post:
      summary: Create exercise plan
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stage:
                  type: string
                  enum: [PRENATAL, POSTNATAL]
                title:
                  type: string
                description:
                  type: string
                duration:
                  type: integer
                source:
                  type: string
              required: [stage, title, description, duration]
      responses:
        '201':
          description: Created plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExercisePlan'
...